+++ src/components/InputField/EXPNumberField.tsx
@@ -1,9 +1,9 @@
 import NumberInput from "../Input/NumberInput";
 import InputGroup from "./InputGroup";
-import { getEXPNumberErrorMessage } from "../../utils/getEXPNumberErrorMessage";
-import { useFieldErrors } from "../../hooks/useFieldErrors";
+import useEXPNumberField, {
+  type EXPNumber,
+} from "../../hooks/useEXPNumberField";
 
-type EXPNumber = { mm: string; yy: string };
 interface Props {
   setEXPNumber: (value: EXPNumber) => void;
   value: EXPNumber;
@@ -15,47 +15,26 @@ export default function EXPNumberField({
   value,
   onComplete,
 }: Props) {
-  const { inputErrors, setError, errorMessage } = useFieldErrors(["mm", "yy"]);
-
-  const handelOnChange = (expKey: string) => (newValue: string) => {
-    setError(expKey as keyof EXPNumber)(null);
-    const newEXPNumbers = { ...value, [expKey]: newValue };
-    setEXPNumber(newEXPNumbers);
-    onComplete(
-      Object.values(newEXPNumbers).every((value) => value.length === 2) &&
-        Object.values(inputErrors).every((err) => err === null),
-    );
-
-    if (expKey === "mm" && newValue.length === 2) {
-      const result = getEXPNumberErrorMessage({ mm: newValue, yy: value.yy });
-      if (result === null || result.key !== "mm") {
-        if (value.yy === "") inputRefs.current["yy"]?.focus();
-      } else setError("mm")(result?.message);
-    }
-  };
-
-  const handleOnBlur =
-    (expKey: string) => (e: React.FocusEvent<HTMLInputElement>) => {
-      const result = getEXPNumberErrorMessage({
-        ...value,
-        [expKey]: e.target.value,
-      });
-      const message = result && result.key === expKey ? result.message : null;
-      setError(expKey as keyof EXPNumber)(message);
-    };
-
+  const {
+    inputRefs,
+    inputErrors,
+    setError,
+    errorMessage,
+    handleOnChange,
+    handleOnBlur,
+  } = useEXPNumberField(setEXPNumber, value, onComplete);
   return (
     <InputGroup errorMessage={errorMessage}>
       {Object.entries(value).map(([expKey, expValue]) => (
         <NumberInput
           key={`${expKey}-input`}
           value={expValue}
-          onChange={handelOnChange(expKey)}
+          onChange={handleOnChange(expKey)}
           placeholder={expKey === "mm" ? "MM" : "YY"}
           autoFocus={expKey === "mm"}
-          hasError={inputErrors[expKey as keyof EXPNumber] !== null}
+          hasError={inputErrors[expKey] !== null}
           maxLength={2}
-          onError={setError(expKey as keyof EXPNumber)}
+          onError={setError(expKey)}
           onBlur={handleOnBlur(expKey)}
           ref={(el) => {
             if (expKey !== "mm") inputRefs.current[expKey] = el;
